Move stream fetch DB queries to worker stores.
